Stop scrolling towards a carousel on load
The .scrollIntoView() DOM method used to implement scrolling via the
buttons is run everytime currentIndex changes. This includes the first
render. This was causing the view to scroll down towards the carousel
whenever the page loaded. The guard clause added here checks whether the
current render is the first render.

The workaround done here will only work on the production build or on
the dev build with React.StrictMode removed. This is because React
intentionally runs two initial renders in the dev build.